refactor MovieController tests

+++ src/main/java/com/example/moviemanager/controller/dto/MovieResponse.java
@@ -2,10 +2,14 @@ package com.example.moviemanager.controller.dto;
 
 import com.example.moviemanager.repository.model.Movie;
 
-public record MovieResponse(Integer id, String title, String description, int releaseYear, double rating) {
+public record MovieResponse(long id, String title, String description, int releaseYear, double rating) {
 
     public static MovieResponse fromEntity(Movie movie) {
-        return new MovieResponse(movie.getId(), movie.getTitle(), movie.getDescription(), movie.getReleaseYear(),
+        return new MovieResponse(
+                movie.getId(),
+                movie.getTitle(),
+                movie.getDescription(),
+                movie.getReleaseYear(),
                 movie.getRating());
     }
 }
