+++ src/test/java/com/example/moviemanager/service/MovieServiceTest.java
@@ -36,8 +36,8 @@ public class MovieServiceTest {
     void returns_list_of_all_movies() {
         // given
         List<Movie> expectedMovies = List.of(
-                new Movie(1, "Home Alone", "Christmas movie", 1990, 8.5),
-                new Movie(2, "Home Alone 2", "Christmas movie", 1992, 8.9));
+                new Movie(1L, "Home Alone", "Christmas movie", 1990, 8.5),
+                new Movie(2L, "Home Alone 2", "Christmas movie", 1992, 8.9));
         when(movieRepository.findAll()).thenReturn(expectedMovies);
 
         // when
@@ -50,7 +50,7 @@ public class MovieServiceTest {
     @Test
     void throws_exception_when_trying_to_find_non_existing_movie() {
         // given
-        Integer movieId = 2;
+        long movieId = 2;
         when(movieRepository.findById(movieId)).thenReturn(Optional.empty());
 
         // then
@@ -62,8 +62,9 @@ public class MovieServiceTest {
     @Test
     void finds_movie_by_id() {
         // given
-        Integer movieId = 1;
-        Movie expectedMovie = new Movie(movieId, "Home Alone", "Christmas movie", 1990, 8.5);
+        long movieId = 1;
+        Movie expectedMovie =
+                new Movie(movieId, "Home Alone", "Christmas movie", 1990, 8.5);
         when(movieRepository.findById(movieId)).thenReturn(Optional.of(expectedMovie));
 
         // when
@@ -76,7 +77,7 @@ public class MovieServiceTest {
     @Test
     void throws_exception_when_trying_to_delete_non_existing_movie() {
         // given
-        Integer movieId = 2;
+        long movieId = 2;
         doThrow(EmptyResultDataAccessException.class).when(movieRepository).deleteById(movieId);
 
         // then
@@ -88,7 +89,7 @@ public class MovieServiceTest {
     @Test
     void deletes_movie_by_id() {
         // given
-        Integer movieId = 1;
+        long movieId = 1;
         doNothing().when(movieRepository).deleteById(movieId);
 
         // when
@@ -102,7 +103,8 @@ public class MovieServiceTest {
     @Test
     void saves_movie() {
         // given
-        Movie expectedMovie = new Movie(1, "Home Alone", "Christmas movie", 1990, 8.5);
+        Movie expectedMovie =
+                new Movie(1L, "Home Alone", "Christmas movie", 1990, 8.5);
         when(movieRepository.save(expectedMovie)).thenReturn(expectedMovie);
 
         // when
@@ -115,11 +117,11 @@ public class MovieServiceTest {
     @Test
     void updates_movie_by_id_with_provided_data() {
         // given
-        Integer movieId = 1;
-        when(movieRepository.findById(movieId))
-                .thenReturn(Optional.of(new Movie(movieId, "Home Alone 2", "Comedy", 1992, 9.0)));
-
-        Movie expectedMovie = new Movie(5, "Home Alone", "Christmas movie", 1990, 8.5);
+        long movieId = 1;
+        when(movieRepository.findById(movieId)).thenReturn(
+                Optional.of(new Movie(movieId, "Home Alone 2", "Comedy", 1992, 9.0)));
+        Movie expectedMovie =
+                new Movie(5L, "Home Alone", "Christmas movie", 1990, 8.5);
         expectedMovie.setId(movieId);
         when(movieRepository.save(expectedMovie)).thenReturn(expectedMovie);
 
@@ -133,8 +135,8 @@ public class MovieServiceTest {
     @Test
     void throws_exception_when_trying_to_update_non_existing_movie() {
         // given
-        Integer movieId = 5;
-        Movie movie = new Movie(1, "Home Alone", "Christmas movie", 1990, 8.5);
+        long movieId = 5;
+        Movie movie = new Movie(1L, "Home Alone", "Christmas movie", 1990, 8.5);
         when(movieRepository.findById(movieId)).thenReturn(Optional.empty());
 
         // then
